Implements acyclic constraint on relationship definitions
This commit adds the acyclic option when defining a directed relationsip
for a StructuredNode class. When the acyclic property is set to True for
a relationship, the RelationshipManager class will do some extra checking
when connect() is called to make sure that if the new relationship were to
be added, it would not create a cycle in the graph for that specific
relationship.